Fix manifest 404 error - remove duplicate manifest link
The vite-plugin-pwa automatically adds the manifest link with the
correct relative path (./manifest.webmanifest), so we don't need
to manually add it in the HTML. The manual link was using an
absolute path (/manifest.webmanifest) which caused 404 errors.

This fixes the issue where the manifest couldn't be loaded,
preventing the PWA from being installable.